moved cache helper
